requirements.txt
Jinja2==3.1.4
Markdown==3.6
PyYAML==6.0.1